own files for eventhandling and AST class
moved rest of ui to corresponding ui file.

forcing most logger prints to when plugin.debug is True
(Need a tickbox so that the user can switch it at their will)
