+++ ui.py
@@ -1,5 +1,17 @@
+"""AST UI functions"""
+
+import logging
+import json
+import os
+from config import appname  # type: ignore
+
 import tkinter as tk
-import myNotebook as nb  # type: ignore # noqa: N813
+import myNotebook as nb  # type: ignore
+from theme import theme  # type: ignore
+from ttkHyperlinkLabel import HyperlinkLabel  # type: ignore
+
+
+logger = logging.getLogger(f"{appname}.{os.path.basename(os.path.dirname(__file__))}")
 
 
 def prefs_label(frame, text, row: int, col: int, sticky) -> None:
@@ -45,3 +57,225 @@ def colourentry(frame, textvariable, row: int, col: int, colour: str, sticky) ->
 def button(frame, text, command, row: int, col: int, sticky) -> None:
     """Create a button for the ui of the plugin."""
     tk.Button(frame, text=text, command=command).grid(row=row, column=col, sticky=sticky)
+
+
+def clear_ui(frame) -> None:
+    """Remove all labels from this plugin."""
+    # remove all labels from the frame
+    for label in frame.winfo_children():
+        label.destroy()
+
+
+def rebuild_ui(plugin, cmdr: str) -> None:  # noqa #CCR001
+    """Rebuild the UI in case of preferences change."""
+
+    clear_ui(plugin.frame)
+
+    # recreate UI
+    current_row = 0
+
+    if plugin.updateavailable:
+        latest = f"github.com/{plugin.AST_REPO}/releases/latest"
+        HyperlinkLabel(plugin.frame, text="Update available!",
+                       url=latest, underline=True).grid(row=current_row, sticky=tk.W)
+        current_row += 1
+
+    uielementcheck = [plugin.AST_hide_fullscan.get(), plugin.AST_hide_species.get(), plugin.AST_hide_progress.get(),
+                      plugin.AST_hide_last_system.get(), plugin.AST_hide_last_body.get(), plugin.AST_hide_value.get(),
+                      plugin.AST_hide_system.get(), plugin.AST_hide_body.get()]
+    uielementlistleft = ["Last Exobiology Scan:", "Last Species:", "Scan Progress:",
+                         "System of last Scan:", "Body of last Scan:", "Unsold Scan Value:",
+                         "Current System:", "Current Body:"]
+    uielementlistright = [plugin.AST_state, plugin.AST_last_scan_plant, plugin.AST_current_scan_progress,
+                          plugin.AST_last_scan_system, plugin.AST_last_scan_body, plugin.AST_value,
+                          plugin.AST_current_system, plugin.AST_current_body]
+    uielementlistextra = [None, None, None, None, None, "clipboardbutton", None, None]
+
+    skipafterselling = ["Last Exobiology Scan:", "Last Species:", "Scan Progress:",
+                        "System of last Scan:", "Body of last Scan:"]
+
+    for i in range(max(len(uielementlistleft), len(uielementlistright))):
+        if uielementcheck[i] != 1:
+            if plugin.AST_after_selling.get() != 0:
+                if uielementlistleft[i] in skipafterselling:
+                    continue
+            # Check when we hide the value of unsold scans when it is 0
+            if uielementlistleft[i] == "Unsold Scan Value:":
+                if (plugin.AST_hide_value_when_zero.get() == 1
+                   and int(plugin.rawvalue) == 0):
+                    continue
+            # Hide when system is the same as the current one.
+            if (uielementlistleft[i] in ["System of last Scan:", "Body of last Scan:"]
+               and (plugin.AST_hide_after_selling.get() == 1 or plugin.AST_hide_after_full_scan.get() == 1)):
+                if uielementlistright[i].get() == uielementlistright[i+3].get():
+                    continue
+            if i < len(uielementlistleft):
+                label(plugin.frame, uielementlistleft[i], current_row, 0, tk.W)
+            if i < len(uielementlistright):
+                entry(plugin.frame, uielementlistright[i], current_row, 1, tk.W)
+            if uielementlistextra[i] == "clipboardbutton":
+                button(plugin.frame, "📋", plugin.clipboard, current_row, 2, tk.E)
+            current_row += 1
+
+    # Clonal Colonial Range here.
+    if plugin.AST_hide_CCR.get() != 1 and plugin.AST_near_planet is True:
+        # show distances for the last scans.
+        colour = "red"
+        if plugin.AST_current_scan_progress.get() in ["0/3", "3/3"]:
+            colour = None
+        if plugin.AST_scan_1_dist_green:
+            colour = "green"
+        colourlabel(plugin.frame, "Distance to Scan #1: ", current_row, 0, colour, tk.W)
+        colourentry(plugin.frame, plugin.AST_scan_1_pos_dist, current_row, 1, colour, tk.W)
+        current_row += 1
+        colour = "red"
+        if plugin.AST_current_scan_progress.get() in ["0/3", "1/3", "3/3"]:
+            colour = None
+        if plugin.AST_scan_2_dist_green:
+            colour = "green"
+        colourlabel(plugin.frame, "Distance to Scan #2: ", current_row, 0, colour, tk.W)
+        colourentry(plugin.frame, plugin.AST_scan_2_pos_dist, current_row, 1, colour, tk.W)
+        current_row += 1
+        colour = None
+        if ((plugin.AST_scan_1_dist_green
+             and plugin.AST_current_scan_progress.get() == "1/3")
+            or (plugin.AST_scan_1_dist_green
+                and plugin.AST_scan_2_dist_green
+                and plugin.AST_current_scan_progress.get() == "2/3")):
+            colour = "green"
+        colourlabel(plugin.frame, "Current Position: ", current_row, 0, colour, tk.W)
+        colourentry(plugin.frame, plugin.AST_current_pos, current_row, 1, colour, tk.W)
+        current_row += 1
+
+    # Tracked sold bio scans as the last thing to add to the UI
+    if plugin.AST_hide_sold_bio.get() != 1:
+        build_sold_bio_ui(plugin, cmdr, current_row)
+
+    theme.update(plugin.frame)  # Apply theme colours to the frame and its children, including the new widgets
+
+
+def build_sold_bio_ui(plugin, cmdr: str, current_row) -> None:  # noqa #CCR001
+    # Create a Button to make it shorter?
+    soldbiodata = {}
+    notsoldbiodata = {}
+
+    file = plugin.AST_DIR + "\\soldbiodata.json"
+    with open(file, "r+", encoding="utf8") as f:
+        soldbiodata = json.load(f)
+
+    file = plugin.AST_DIR + "\\notsoldbiodata.json"
+    with open(file, "r+", encoding="utf8") as f:
+        notsoldbiodata = json.load(f)
+
+    label(plugin.frame, "Scans in this System:", current_row, 0, tk.W)
+
+    if cmdr == "" or cmdr is None or cmdr == "None":
+        return
+
+    # Check if we even got a cmdr yet!
+    if plugin.debug:
+        logger.info(f"Commander: {cmdr}. attempting to access")
+        logger.info(f"data: {soldbiodata[cmdr]}.")
+        logger.info(f"data: {notsoldbiodata}.")
+
+    bodylistofspecies = {}
+    try:
+        firstletter = plugin.AST_current_system.get()[0].lower()
+    except IndexError:
+        label(plugin.frame, "None", current_row, 1, tk.W)
+        # length of string is 0. there is no current system yet.
+        # So there is no reason to do anything
+        return
+
+    count = 0
+    count_from_planet = 0
+    currentbody = plugin.AST_current_body.get().replace(plugin.AST_current_system.get(), "")[1:]
+    if plugin.debug:
+        logger.debug(plugin.AST_num_bios_on_planet)
+
+    try:
+        if plugin.AST_current_system.get() in soldbiodata[cmdr][firstletter].keys():
+            for sold in soldbiodata[cmdr][firstletter][plugin.AST_current_system.get()]:
+                bodyname = ""
+
+                # Check if body has a special name or if we have standardized names
+                if sold["system"] in sold["body"]:
+                    # no special name for planet
+                    bodyname = sold["body"].replace(sold["system"], "")[1:]
+                else:
+                    bodyname = sold["body"]
+
+                if sold["species"] not in bodylistofspecies.keys():
+                    bodylistofspecies[sold["species"]] = [[bodyname, True]]
+                else:
+                    bodylistofspecies[sold["species"]].append([bodyname, True])
+
+                if plugin.debug:
+                    logger.debug(f"{bodyname} checked and this is the current: {currentbody}")
+
+                if currentbody == bodyname:
+                    count_from_planet += 1
+                count += 1
+    except KeyError:
+        # if we don't have the cmdr in the sold data yet we just pass all sold data.
+        pass
+
+    try:
+        for notsold in notsoldbiodata[cmdr]:
+            if notsold["system"] == plugin.AST_current_system.get():
+
+                bodyname = ""
+
+                # Check if body has a special name or if we have standardized names
+                if notsold["system"] in notsold["body"]:
+                    # no special name for planet
+                    bodyname = notsold["body"].replace(notsold["system"], "")[1:]
+                else:
+                    bodyname = notsold["body"]
+
+                if notsold["species"] not in bodylistofspecies.keys():
+                    bodylistofspecies[notsold["species"]] = [[bodyname, False]]
+                else:
+                    bodylistofspecies[notsold["species"]].append([bodyname, False])
+
+                if plugin.debug:
+                    logger.debug(f"{bodyname} checked and this is the current: {currentbody}")
+
+                if currentbody == bodyname:
+                    count_from_planet += 1
+                count += 1
+    except KeyError:
+        # if we don't have the cmdr in the notsold data yet we just pass.
+        pass
+
+    if bodylistofspecies == {}:
+        label(plugin.frame, "None", current_row, 1, tk.W)
+    else:
+        label(plugin.frame, count, current_row, 1, tk.W)
+
+    if plugin.AST_num_bios_on_planet != 0:
+        label(plugin.frame, "on Planet:", current_row, 2, tk.W)
+        amount_found_of_total = f"{count_from_planet}/{plugin.AST_num_bios_on_planet}"
+        label(plugin.frame, amount_found_of_total, current_row, 3, tk.W)
+
+    # skip
+    if plugin.AST_hide_scans_in_system.get() != 0:
+        button(plugin.frame, "▼", plugin.switchhidesoldexobio, current_row, 4, tk.W)
+
+        return
+
+    button(plugin.frame, "▲", plugin.switchhidesoldexobio, current_row, 4, tk.W)
+
+    for species in bodylistofspecies.keys():
+        current_row += 1
+        label(plugin.frame, species, current_row, 0, tk.W)
+        bodies = ""
+        for body in bodylistofspecies[species]:
+            if body[1]:
+                bodies = bodies + body[0] + ", "
+            else:
+                bodies = bodies + "*" + body[0] + "*, "
+        while (bodies[-1] == "," or bodies[-1] == " "):
+            bodies = bodies[:-1]
+
+        label(plugin.frame, bodies, current_row, 1, tk.W)
